Reorganize C++ half significantly
