+++ cpp/bindings/types.h
